+++ packages/back-end/src/integrations/SqlIntegration.ts
@@ -37,7 +37,6 @@ import {
   isMultiStatementSQL,
   SQL_ROW_LIMIT,
 } from "shared/sql";
-import { MissingDatasourceParamsError } from "shared/util";
 import {
   PhaseSQLVar,
   SQLVars,
@@ -142,6 +141,7 @@ import {
 } from "shared/types/fact-table";
 import type { PopulationDataQuerySettings } from "shared/types/query";
 import { AdditionalQueryMetadata, QueryMetadata } from "shared/types/query";
+import { MissingDatasourceParamsError } from "back-end/src/util/errors";
 import { UNITS_TABLE_PREFIX } from "back-end/src/queryRunners/ExperimentResultsQueryRunner";
 import { ReqContext } from "back-end/types/request";
 import { SourceIntegrationInterface } from "back-end/src/types/Integration";
