+++ packages/back-end/src/jobs/createInformationSchema.ts
@@ -3,7 +3,7 @@ import { InformationSchemaError } from "shared/types/integrations";
 import {
   DataSourceNotSupportedError,
   MissingDatasourceParamsError,
-} from "shared/util";
+} from "back-end/src/util/errors";
 import { getDataSourceById } from "back-end/src/models/DataSourceModel";
 import {
   getInformationSchemaByDatasourceId,
